+++ example/factories.ts
@@ -98,3 +98,15 @@ export const statusDependingFactory = new EndpointsFactory({
     },
   }),
 });
+
+/** @desc This factory demonstrates response without body, such as 204 No Content */
+export const noContentFactory = new EndpointsFactory({
+  config,
+  resultHandler: new ResultHandler({
+    positive: { statusCode: 204, mimeType: null, schema: z.never() },
+    negative: { statusCode: 404, mimeType: null, schema: z.never() },
+    handler: ({ error, response }) => {
+      response.status(error ? ensureHttpError(error).statusCode : 204).end(); // no content
+    },
+  }),
+});