+++ src/api-response.ts
@@ -7,19 +7,28 @@ export const defaultStatusCodes = {
 
 export type ResponseVariant = keyof typeof defaultStatusCodes;
 
+/** @public this is the user facing configuration */
 export interface ApiResponse<S extends z.ZodTypeAny> {
   schema: S;
   /** @default 200 for a positive and 400 for a negative response */
   statusCode?: number | [number, ...number[]];
-  /** @default "application/json" */
-  mimeType?: string | [string, ...string[]];
+  /**
+   * @example null is for no content, such as 204 and 302
+   * @default "application/json"
+   * */
+  mimeType?: string | [string, ...string[]] | null;
   /** @deprecated use statusCode */
   statusCodes?: never;
   /** @deprecated use mimeType */
   mimeTypes?: never;
 }
 
-export type NormalizedResponse = Pick<ApiResponse<z.ZodTypeAny>, "schema"> & {
-  statusCodes: Extract<ApiResponse<z.ZodTypeAny>["statusCode"], Array<unknown>>;
-  mimeTypes: Extract<ApiResponse<z.ZodTypeAny>["mimeType"], Array<unknown>>;
-};
+/**
+ * @private This is what the framework entities operate
+ * @see normalize
+ * */
+export interface NormalizedResponse {
+  schema: z.ZodTypeAny;
+  statusCodes: [number, ...number[]];
+  mimeTypes: [string, ...string[]] | null;
+}