+++ src/endpoint.ts
@@ -20,7 +20,7 @@ import { ActualLogger } from "./logger-helpers";
 import { LogicalContainer, combineContainers } from "./logical-container";
 import { AuxMethod, Method } from "./method";
 import { AbstractMiddleware, ExpressMiddleware } from "./middleware";
-import { ContentType, contentTypes } from "./content-type";
+import { ContentType } from "./content-type";
 import { Nesting } from "./nesting";
 import { AbstractResultHandler } from "./result-handler";
 import { Security } from "./security";
@@ -33,7 +33,6 @@ export type Handler<IN, OUT, OPT> = (params: {
 
 type DescriptionVariant = "short" | "long";
 type IOVariant = "input" | "output";
-type MimeVariant = Extract<IOVariant, "input"> | ResponseVariant;
 
 export abstract class AbstractEndpoint extends Nesting {
   public abstract execute(params: {
@@ -47,8 +46,6 @@ export abstract class AbstractEndpoint extends Nesting {
   ): string | undefined;
   public abstract getMethods(): ReadonlyArray<Method> | undefined;
   public abstract getSchema(variant: IOVariant): IOSchema;
-  public abstract getSchema(variant: ResponseVariant): z.ZodTypeAny;
-  public abstract getMimeTypes(variant: MimeVariant): ReadonlyArray<string>;
   public abstract getResponses(
     variant: ResponseVariant,
   ): ReadonlyArray<NormalizedResponse>;
@@ -69,7 +66,6 @@ export class Endpoint<
   readonly #descriptions: Record<DescriptionVariant, string | undefined>;
   readonly #methods?: ReadonlyArray<Method>;
   readonly #middlewares: AbstractMiddleware[];
-  readonly #mimeTypes: Record<MimeVariant, ReadonlyArray<string>>;
   readonly #responses: Record<
     ResponseVariant,
     ReadonlyArray<NormalizedResponse>
@@ -126,15 +122,6 @@ export class Endpoint<
       : hasRaw(inputSchema)
         ? "raw"
         : "json";
-    this.#mimeTypes = {
-      input: Object.freeze([contentTypes[this.#requestType]]),
-      positive: Object.freeze(
-        this.#responses.positive.flatMap(({ mimeTypes }) => mimeTypes),
-      ),
-      negative: Object.freeze(
-        this.#responses.negative.flatMap(({ mimeTypes }) => mimeTypes),
-      ),
-    };
   }
 
   public override getDescription(variant: DescriptionVariant) {
@@ -147,17 +134,8 @@ export class Endpoint<
 
   public override getSchema(variant: "input"): IN;
   public override getSchema(variant: "output"): OUT;
-  public override getSchema(variant: ResponseVariant): z.ZodTypeAny;
-  public override getSchema(variant: IOVariant | ResponseVariant) {
-    if (variant === "input" || variant === "output")
-      return this.#schemas[variant];
-    return this.getResponses(variant)
-      .map(({ schema }) => schema)
-      .reduce((agg, schema) => agg.or(schema));
-  }
-
-  public override getMimeTypes(variant: MimeVariant) {
-    return this.#mimeTypes[variant];
+  public override getSchema(variant: IOVariant) {
+    return this.#schemas[variant];
   }
 
   public override getRequestType() {